+++ src/app/modules/stock/components/stock-analyse/stock-analyse.component.scss
