+++ src/app/modules/stock/components/stock-analyse/stock-analyse.component.spec.ts
@@ -1,18 +1,18 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { StockDetailsComponent } from './stock-details.component';
+import { StockAnalyseComponent } from './stock-analyse.component';
 
 describe('StockComponent', () => {
-  let component: StockDetailsComponent;
-  let fixture: ComponentFixture<StockDetailsComponent>;
+  let component: StockAnalyseComponent;
+  let fixture: ComponentFixture<StockAnalyseComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [StockDetailsComponent]
+      imports: [StockAnalyseComponent]
     })
     .compileComponents();
 
-    fixture = TestBed.createComponent(StockDetailsComponent);
+    fixture = TestBed.createComponent(StockAnalyseComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
