+++ src/app/modules/stock/components/stock-analyse/stock-analyse.config.ts
@@ -1,6 +1,6 @@
 import { DxButtonTypes } from 'devextreme-angular/ui/button';
 import { Currency } from '../../../profile/components/profile/profile-wallets/profile-wallets.config';
-import { ChartType, CryptoDetails, TimeFrame } from './stock-details.model';
+import { ChartType, CryptoDetails, TimeFrame } from './stock-analyse.model';
 
 export const defaultCrypto: Currency = {
   currency_name: 'bitcoin',
@@ -112,6 +112,6 @@ export const stockQueryParamNames = {
 };
 
 export const defaultChartType: ChartType = {
-  text: $localize`:@@stock-details.Candlestick:Candlestick`,
+  text: $localize`:@@stock-analyse.Candlestick:Candlestick`,
   value: 'candlestick',
 };
