+++ src/app/modules/stock/components/stock-analyse/stock-order/stock-order.component.spec.ts
