+++ src/app/modules/stock/components/stock-analyse/stock-order/stock-order.config.ts
