+++ src/app/modules/stock/stock.module.ts
@@ -15,8 +15,8 @@ import { DxDataGridModule } from 'devextreme-angular/ui/data-grid';
 
 import { routes } from './stock.routes';
 import { HeaderComponent } from '../../shared/components/header/header.component';
-import { StockDetailsComponent } from './components/stock-details/stock-details.component';
-import { StockOrderComponent } from './components/stock-details/stock-order/stock-order.component';
+import { StockAnalyseComponent } from './components/stock-analyse/stock-analyse.component';
+import { StockOrderComponent } from './components/stock-analyse/stock-order/stock-order.component';
 import { StocksListComponent } from './components/stocks-list/stocks-list.component';
 import { StockComponent } from './components/stock/stock.component';
 import { GridDeleteButtonDirective } from '../../directives/grid-delete-button/grid-delete-button.directive';
@@ -27,7 +27,7 @@ import { GridDeleteButtonDirective } from '../../directives/grid-delete-button/g
     provideRouter(routes),
   ],
   declarations: [
-    StockDetailsComponent,
+    StockAnalyseComponent,
     StockOrderComponent,
     StocksListComponent,
     StockComponent,
