+++ src/app/modules/stock/stock.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
 import { authGuard } from '../../guards/auth/auth.guard';
-import { StockDetailsComponent } from './components/stock-details/stock-details.component';
+import { StockAnalyseComponent } from './components/stock-analyse/stock-analyse.component';
 import { StocksListComponent } from './components/stocks-list/stocks-list.component';
 import { StockComponent } from './components/stock/stock.component';
 
@@ -16,7 +16,7 @@ export const routes: Routes = [
       },
       {
         path: 'analyse',
-        component: StockDetailsComponent,
+        component: StockAnalyseComponent,
       },
       {
         path: '**',
